+++ lib/async_events.dart
@@ -1,5 +1,6 @@
 /// Asynchronous event hub library.
 library async_events;
 
+export 'src/async_call.dart';
 export 'src/async_events_base.dart';
 export 'src/async_events_storage.dart';